+++ Locales/enUS.lua
@@ -52,8 +52,12 @@ select(2, ...).L = setmetatable({
     ["RAID_DEBUFFS_TIPS"] = "Tips: [Drag & Drop] to change debuff order. [Double-Click] on instance name to open Encounter Journal. [Shift+Left Click] on instance/boss name to share debuffs. [Alt+Left Click] on instance/boss name to reset debuffs. The priority of General Debuffs is higher than Boss Debuffs.",
 
     ["CHANGE LOGS"] = [[
+        <h1>r86-release (Mar 27, 2022, 15:00 GMT+8)</h1>
+        <p>* Added a "Default" anchor option for tooltips.</p>
+        <br/>
+
         <h1>r85-release (Mar 26, 2022, 18:00 GMT+8)</h1>
-        <p>* Fixed bugs (occured when scale ≠ 1).</p>
+        <p>* Fixed bugs (occured when scale ~= 1).</p>
         <br/>
 
         <h1>r84-release (Mar 26, 2022, 15:45 GMT+8)</h1>